Performance improvements when fetching status items via obj.getChild.
According to Jamie Teh (NV Access), when looking for a specific child, it is faster to use obj.getChild(index) instead of obj.children[index]. Subsequent test shows it is indeed the case - up to four times faster than before. This means performance improvements when fetching information such as channels, selection duration and others.
This is destined for add-on 4.0 (later in summer).