Track remainig time/convert time to seconds: 'times' is now a list of string lists instead of a argument tuple.
Clarity: time string to seconds conversion 'times' argument was originally defined as a dynamic tuple of positional arguments. However, this made code inspection difficult and type annotation could not be defined. Therefore, define this argument as a list of string lists (times) to improve clarity and to add type annotations.